+++ Package.swift
@@ -14,7 +14,7 @@ let package = Package(
         )
     ],
     dependencies: [
-        .package(url: "https://github.com/patrickridd/AuthDomain.git", from: "1.1.0")
+        .package(url: "https://github.com/patrickridd/AuthDomain.git", from: "1.2.0")
     ],
     targets: [
         .target(